[ADD] MVVM 적용(폴더링, ViewModel생성), Hilt 추가
