+++ app/src/main/AndroidManifest.xml
@@ -1,40 +1,24 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.seoulyanghang.android">
+    package="com.seoulyanghang.android" >
 
     <application
+        android:name=".SeoulYanghangApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.SeoulYanghang">
+        android:theme="@style/Theme.SeoulYanghang" >
         <activity
-            android:name=".ui.user.LogInActivity"
+            android:name=".ui.login.view.LogInActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.user.SignInActivity"
-            android:exported="false" />
-        <activity
-            android:name=".ui.user.login.signup.SignUpActivity"
-            android:exported="false"
-            android:label="@string/title_activity_sign_up"
-            android:theme="@style/Theme.SeoulYanghang.NoActionBar" />
-        <activity
-            android:name=".ui.user.MainActivitydddd"
-            android:exported="false" />
-        <activity
-            android:name=".ui.user.signup.SignUpActivity"
-            android:exported="false" />
-        <activity
-            android:name=".ui.user.login.LogInActivity"
-            android:exported="false" />
-        <activity
-            android:name=".ui.MainActivityvx"
+            android:name=".ui.signup.view.SignUpActivity"
             android:exported="false" />
         <activity
             android:name=".ui.MainActivity"
-            android:exported="true">
+            android:exported="true" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
