+++ app/src/main/java/com/seoulyanghang/android/ui/itemreview/view/ItemReviewFragment.kt
@@ -1,19 +1,18 @@
-package com.seoulyanghang.android.ui.loveitem
+package com.seoulyanghang.android.ui.itemreview.view
 
-import android.content.Intent
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.recyclerview.widget.LinearLayoutManager
 import com.seoulyanghang.android.R
 import com.seoulyanghang.android.base.BaseViewUtil
+import com.seoulyanghang.android.databinding.FragmentItemReviewBinding
 import com.seoulyanghang.android.databinding.FragmentLoveItemBinding
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class LoveItemFragment : BaseViewUtil.BaseFragment<FragmentLoveItemBinding>(R.layout.fragment_love_item) {
+class ItemReviewFragment : BaseViewUtil.BaseFragment<FragmentItemReviewBinding>(R.layout.fragment_item_review) {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
