+++ app/src/main/java/com/seoulyanghang/android/ui/loveitem/view/LoveItemFragment.kt
@@ -1,18 +1,19 @@
-package com.seoulyanghang.android.ui.recentitem
+package com.seoulyanghang.android.ui.loveitem.view
 
+import android.content.Intent
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.recyclerview.widget.LinearLayoutManager
 import com.seoulyanghang.android.R
 import com.seoulyanghang.android.base.BaseViewUtil
 import com.seoulyanghang.android.databinding.FragmentLoveItemBinding
-import com.seoulyanghang.android.databinding.FragmentRecentItemBinding
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class RecentItemFragment : BaseViewUtil.BaseFragment<FragmentRecentItemBinding>(R.layout.fragment_recent_item) {
+class LoveItemFragment : BaseViewUtil.BaseFragment<FragmentLoveItemBinding>(R.layout.fragment_love_item) {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
