+++ app/src/main/java/com/seoulyanghang/android/ui/loveitem/viewmodel/LoveItemViewModel.kt
@@ -1,0 +1,9 @@
+package com.seoulyanghang.android.ui.loveitem.viewmodel
+
+import androidx.lifecycle.ViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+
+@HiltViewModel
+class LoveItemViewModel @Inject constructor() : ViewModel() {
+}
