+++ app/src/main/java/com/seoulyanghang/android/ui/recentitem/view/RecentItemFragment.kt
@@ -1,4 +1,4 @@
-package com.seoulyanghang.android.ui.itemreview
+package com.seoulyanghang.android.ui.recentitem.view
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,12 +7,12 @@ import android.view.View
 import android.view.ViewGroup
 import com.seoulyanghang.android.R
 import com.seoulyanghang.android.base.BaseViewUtil
-import com.seoulyanghang.android.databinding.FragmentItemReviewBinding
 import com.seoulyanghang.android.databinding.FragmentLoveItemBinding
+import com.seoulyanghang.android.databinding.FragmentRecentItemBinding
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class ItemReviewFragment : BaseViewUtil.BaseFragment<FragmentItemReviewBinding>(R.layout.fragment_item_review) {
+class RecentItemFragment : BaseViewUtil.BaseFragment<FragmentRecentItemBinding>(R.layout.fragment_recent_item) {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
