+++ app/src/main/java/com/seoulyanghang/android/ui/recentitem/viewmodel/RecentItemViewModel.kt
@@ -1,0 +1,9 @@
+package com.seoulyanghang.android.ui.recentitem.viewmodel
+
+import androidx.lifecycle.ViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+
+@HiltViewModel
+class RecentItemViewModel @Inject constructor() : ViewModel() {
+}
