+++ app/src/main/java/com/seoulyanghang/android/ui/signup/viewmodel/SignUpViewModel.kt
@@ -1,0 +1,9 @@
+package com.seoulyanghang.android.ui.signup.viewmodel
+
+import androidx.lifecycle.ViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+
+@HiltViewModel
+class SignUpViewModel @Inject constructor(): ViewModel(){
+}
