--- app/src/main/java/com/seoulyanghang/android/ui/user/LogInActivity.kt
@@ -1,13 +1,0 @@
-package com.seoulyanghang.android.ui.user
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import com.seoulyanghang.android.R
-import com.seoulyanghang.android.base.BaseViewUtil
-import com.seoulyanghang.android.databinding.ActivityLogInBinding
-
-class LogInActivity : BaseViewUtil.BaseAppCompatActivity<ActivityLogInBinding>(R.layout.activity_log_in) {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-    }
-}
