+++ app/src/main/res/layout/activity_log_in.xml
@@ -3,7 +3,6 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools">
 
-
     <data>
 
     </data>
@@ -11,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.user.LogInActivity">
+        tools:context=".ui.login.view.LogInActivity">
 
     </androidx.constraintlayout.widget.ConstraintLayout>
