+++ app/src/main/res/layout/fragment_item_review.xml
@@ -9,7 +9,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.itemreview.ItemReviewFragment">
+        tools:context=".ui.itemreview.view.ItemReviewFragment">
 
         <TextView
             android:layout_width="match_parent"
