+++ app/src/main/res/layout/fragment_love_item.xml
@@ -9,7 +9,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.loveitem.LoveItemFragment">
+        tools:context=".ui.loveitem.view.LoveItemFragment">
 
         <TextView
             android:layout_width="match_parent"
