+++ app/src/main/res/layout/fragment_recent_item.xml
@@ -9,7 +9,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.recentitem.RecentItemFragment">
+        tools:context=".ui.recentitem.view.RecentItemFragment">
 
         <TextView
             android:layout_width="match_parent"
