+++ app/src/main/res/layout/fragment_shopping_basket.xml
@@ -9,7 +9,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.shoppingbasket.ShoppingBasketFragment">
+        tools:context=".ui.shoppingbasket.view.ShoppingBasketFragment">
 
 
         <TextView
