+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/FirstFragment"
-        android:name="com.seoulyanghang.android.ui.user.login.signup.FirstFragment"
+        android:name="com.seoulyanghang.android.ui.login.login.signup.FirstFragment"
         android:label="@string/first_fragment_label"
         tools:layout="@layout/fragment_first">
 
@@ -17,7 +17,7 @@
     </fragment>
     <fragment
         android:id="@+id/SecondFragment"
-        android:name="com.seoulyanghang.android.ui.user.login.signup.SecondFragment"
+        android:name="com.seoulyanghang.android.ui.login.login.signup.SecondFragment"
         android:label="@string/second_fragment_label"
         tools:layout="@layout/fragment_second">
